+++ src/marionette.rs
@@ -40,11 +40,10 @@ use webdriver::command::{
     SwitchToFrameParameters, LocatorParameters, JavascriptCommandParameters,
     GetNamedCookieParameters, AddCookieParameters, TimeoutsParameters,
     ActionsParameters, TakeScreenshotParameters};
-use webdriver::response::{CloseWindowResponse, Cookie, CookieResponse, ElementRectResponse,
-                          NewSessionResponse, TimeoutsResponse, ValueResponse, WebDriverResponse,
-                          WindowRectResponse};
-use webdriver::common::{
-    Date, Nullable, WebElement, FrameId, ELEMENT_KEY};
+use webdriver::response::{CloseWindowResponse, Cookie, CookieResponse, CookiesResponse,
+                          NewSessionResponse, RectResponse, TimeoutsResponse, ValueResponse,
+                          WebDriverResponse};
+use webdriver::common::{Date, ELEMENT_KEY, FrameId, Nullable, WebElement};
 use webdriver::error::{ErrorStatus, WebDriverError, WebDriverResult};
 use webdriver::server::{WebDriverHandler, Session};
 use webdriver::httpapi::{WebDriverExtensionRoute};
@@ -637,7 +636,7 @@ impl MarionetteSession {
         Ok(match msg.command {
             // Everything that doesn't have a response value
             Get(_) | GoBack | GoForward | Refresh | SetTimeouts(_) |
-            MaximizeWindow | SwitchToWindow(_) | SwitchToFrame(_) |
+            SwitchToWindow(_) | SwitchToFrame(_) |
             SwitchToParentFrame | AddCookie(_) | DeleteCookies | DeleteCookie(_) |
             DismissAlert | AcceptAlert | SendAlertText(_) | ElementClick(_) |
             ElementTap(_) | ElementClear(_) | ElementSendKeys(_, _) |
@@ -704,40 +703,6 @@ impl MarionetteSession {
                                             })
                                        .collect());
                 WebDriverResponse::CloseWindow(CloseWindowResponse { window_handles: handles })
-            }
-            GetWindowRect => {
-                let width = try_opt!(
-                    try_opt!(resp.result.find("width"),
-                             ErrorStatus::UnknownError,
-                             "Failed to find width field").as_u64(),
-                    ErrorStatus::UnknownError,
-                    "Failed to interpret width as integer");
-
-                let height = try_opt!(
-                    try_opt!(resp.result.find("height"),
-                             ErrorStatus::UnknownError,
-                             "Failed to find height field").as_u64(),
-                    ErrorStatus::UnknownError,
-                    "Failed to interpret width as integer");
-
-                let x = try_opt!(
-                    try_opt!(resp.result.find("x"),
-                             ErrorStatus::UnknownError,
-                             "Failed to find x field").as_i64(),
-                    ErrorStatus::UnknownError,
-                    "Failed to interpret x as integer");
-
-                let y = try_opt!(
-                    try_opt!(resp.result.find("y"),
-                             ErrorStatus::UnknownError,
-                             "Failed to find y field").as_i64(),
-                    ErrorStatus::UnknownError,
-                    "Failed to interpret y as integer");
-
-                WebDriverResponse::WindowRect(WindowRectResponse {x: x,
-                                                                  y: y,
-                                                                  width: width,
-                                                                  height: height})
             },
             GetElementRect(_) => {
                 let x = try_opt!(
@@ -768,9 +733,24 @@ impl MarionetteSession {
                     ErrorStatus::UnknownError,
                     "Failed to interpret width as float");
 
-                WebDriverResponse::ElementRect(ElementRectResponse::new(x, y, width, height))
+                WebDriverResponse::ElementRect(RectResponse::new(x, y, width, height))
             },
+            FullscreenWindow | MaximizeWindow | GetWindowRect |
             SetWindowRect(_) => {
+                let width = try_opt!(
+                    try_opt!(resp.result.find("width"),
+                             ErrorStatus::UnknownError,
+                             "Failed to find width field").as_f64(),
+                    ErrorStatus::UnknownError,
+                    "Failed to interpret width as float");
+
+                let height = try_opt!(
+                    try_opt!(resp.result.find("height"),
+                             ErrorStatus::UnknownError,
+                             "Failed to find height field").as_f64(),
+                    ErrorStatus::UnknownError,
+                    "Failed to interpret height as float");
+
                 let x = try_opt!(
                     try_opt!(resp.result.find("x"),
                              ErrorStatus::UnknownError,
@@ -785,64 +765,19 @@ impl MarionetteSession {
                     ErrorStatus::UnknownError,
                     "Failed to interpret y as float");
 
-                let width = try_opt!(
-                    try_opt!(resp.result.find("width"),
-                             ErrorStatus::UnknownError,
-                             "Failed to find width field").as_f64(),
-                    ErrorStatus::UnknownError,
-                    "Failed to interpret width as float");
-
-                let height = try_opt!(
-                    try_opt!(resp.result.find("height"),
-                             ErrorStatus::UnknownError,
-                             "Failed to find height field").as_f64(),
-                    ErrorStatus::UnknownError,
-                    "Failed to interpret width as float");
-
-                WebDriverResponse::ElementRect(ElementRectResponse::new(x, y, width, height))
-            },
-            FullscreenWindow => {
-                let width = try_opt!(
-                    try_opt!(resp.result.find("width"),
-                             ErrorStatus::UnknownError,
-                             "Failed to find width field").as_u64(),
-                    ErrorStatus::UnknownError,
-                    "Failed to interpret width as integer");
-
-                let height = try_opt!(
-                    try_opt!(resp.result.find("height"),
-                             ErrorStatus::UnknownError,
-                             "Failed to find height field").as_u64(),
-                    ErrorStatus::UnknownError,
-                    "Failed to interpret height as integer");
-
-                let x = try_opt!(
-                    try_opt!(resp.result.find("x"),
-                             ErrorStatus::UnknownError,
-                             "Failed to find x field").as_i64(),
-                    ErrorStatus::UnknownError,
-                    "Failed to interpret x as integer");
-
-                let y = try_opt!(
-                    try_opt!(resp.result.find("y"),
-                             ErrorStatus::UnknownError,
-                             "Failed to find y field").as_i64(),
-                    ErrorStatus::UnknownError,
-                    "Failed to interpret y as integer");
-
-                WebDriverResponse::WindowRect(WindowRectResponse {x: x,
-                                                                  y: y,
-                                                                  width: width,
-                                                                  height: height})
+                WebDriverResponse::WindowRect(RectResponse::new(x, y, width, height))
             },
             GetCookies => {
                 let cookies = try!(self.process_cookies(&resp.result));
-                WebDriverResponse::Cookie(CookieResponse::new(cookies))
+                WebDriverResponse::Cookies(CookiesResponse { value: cookies })
             },
             GetNamedCookie(ref name) => {
                 let mut cookies = try!(self.process_cookies(&resp.result));
                 cookies.retain(|x| x.name == *name);
-                WebDriverResponse::Cookie(CookieResponse::new(cookies))
+                let cookie = try_opt!(cookies.pop(),
+                                      ErrorStatus::NoSuchCookie,
+                                      format!("No cookie with name {}", name));
+                WebDriverResponse::Cookie(CookieResponse { value: cookie })
             }
             FindElement(_) | FindElementElement(_, _) => {
                 let element = try!(self.to_web_element(
@@ -936,28 +871,28 @@ impl MarionetteSession {
             let name = try_opt!(
                 try_opt!(x.find("name"),
                          ErrorStatus::UnknownError,
-                         "Failed to find name field").as_string(),
+                         "Cookie must have a name field").as_string(),
                 ErrorStatus::UnknownError,
-                "Failed to interpret name as string").to_string();
+                "Cookie must have string name").to_string();
             let value = try_opt!(
                 try_opt!(x.find("value"),
                          ErrorStatus::UnknownError,
-                         "Failed to find value field").as_string(),
+                         "Cookie must have a value field").as_string(),
                 ErrorStatus::UnknownError,
-                "Failed to interpret value as string").to_string();
+                "Cookie must have a string value").to_string();
             let path = try!(
                 Nullable::from_json(x.find("path").unwrap_or(&Json::Null),
                                     |x| {
                                         Ok((try_opt!(x.as_string(),
                                                      ErrorStatus::UnknownError,
-                                                     "Failed to interpret path as String")).to_string())
+                                                     "Cookie path must be string")).to_string())
                                     }));
             let domain = try!(
                 Nullable::from_json(x.find("domain").unwrap_or(&Json::Null),
                                     |x| {
                                         Ok((try_opt!(x.as_string(),
                                                      ErrorStatus::UnknownError,
-                                                     "Failed to interpret domain as String")).to_string())
+                                                     "Cookie domain must be string")).to_string())
                                     }));
             let expiry = try!(
                 Nullable::from_json(x.find("expiry").unwrap_or(&Json::Null),
@@ -965,17 +900,27 @@ impl MarionetteSession {
                                         Ok(Date::new((try_opt!(
                                             x.as_u64(),
                                             ErrorStatus::UnknownError,
-                                            "Failed to interpret expiry as u64"))))
+                                            "Cookie expiry must be a positive integer"))))
                                     }));
             let secure = try_opt!(
                 x.find("secure").map_or(Some(false), |x| x.as_boolean()),
                 ErrorStatus::UnknownError,
-                "Failed to interpret secure as boolean");
+                "Cookie secure flag must be boolean");
             let http_only = try_opt!(
                 x.find("httpOnly").map_or(Some(false), |x| x.as_boolean()),
                 ErrorStatus::UnknownError,
-                "Failed to interpret httpOnly as boolean");
-            Ok(Cookie::new(name, value, path, domain, expiry, secure, http_only))
+                "Cookie httpOnly flag must be boolean");
+
+            let new_cookie = Cookie {
+                name: name,
+                value: value,
+                path: path,
+                domain: domain,
+                expiry: expiry,
+                secure: secure,
+                httpOnly: http_only,
+            };
+            Ok(new_cookie)
         }).collect::<Result<Vec<_>, _>>()
     }
 
@@ -1050,7 +995,7 @@ impl MarionetteCommand {
             DeleteSession => {
                 let mut body = BTreeMap::new();
                 body.insert("flags".to_owned(), vec!["eForceQuit".to_json()].to_json());
-                (Some("quitApplication"), Some(Ok(body)))
+                (Some("quit"), Some(Ok(body)))
             },
             Status => panic!("Got status command that should already have been handled"),
             Get(ref x) => (Some("get"), Some(x.to_marionette())),
@@ -1064,7 +1009,7 @@ impl MarionetteCommand {
             GetWindowHandles => (Some("getWindowHandles"), None),
             CloseWindow => (Some("close"), None),
             GetTimeouts => (Some("getTimeouts"), None),
-            SetTimeouts(ref x) => (Some("timeouts"), Some(x.to_marionette())),
+            SetTimeouts(ref x) => (Some("setTimeouts"), Some(x.to_marionette())),
             SetWindowRect(ref x) => (Some("setWindowRect"), Some(x.to_marionette())),
             GetWindowRect => (Some("getWindowRect"), None),
             MaximizeWindow => (Some("maximizeWindow"), None),